fix: move package to eliminate deps
now the /pkg directory does not have any dependencies to the services
