+++ services/cd-service/pkg/argocd/reposerver/reposerver_test.go
@@ -18,6 +18,7 @@ package reposerver
 
 import (
 	"context"
+	"github.com/freiheit-com/kuberpult/services/cd-service/pkg/repository/testutil"
 	"os"
 	"os/exec"
 	"path"
@@ -35,7 +36,6 @@ import (
 	"github.com/argoproj/argo-cd/v2/util/argo"
 	cacheutil "github.com/argoproj/argo-cd/v2/util/cache"
 	"github.com/argoproj/argo-cd/v2/util/git"
-	"github.com/freiheit-com/kuberpult/pkg/testutil"
 	"github.com/freiheit-com/kuberpult/services/cd-service/pkg/config"
 	"github.com/freiheit-com/kuberpult/services/cd-service/pkg/repository"
 	"github.com/google/go-cmp/cmp"
