+++ services/cd-service/pkg/service/overview_test.go
@@ -18,13 +18,12 @@ package service
 
 import (
 	"context"
+	"github.com/freiheit-com/kuberpult/services/cd-service/pkg/repository/testutil"
 	"sync"
 	"testing"
 
-	"github.com/freiheit-com/kuberpult/pkg/auth"
-	"github.com/freiheit-com/kuberpult/pkg/testutil"
-
 	api "github.com/freiheit-com/kuberpult/pkg/api/v1"
+	"github.com/freiheit-com/kuberpult/pkg/auth"
 	"github.com/freiheit-com/kuberpult/services/cd-service/pkg/config"
 	"github.com/freiheit-com/kuberpult/services/cd-service/pkg/repository"
 	"github.com/google/go-cmp/cmp"
